Actually get things building

+++ wpical/src/main/native/cpp/cameracalibration.cpp
@@ -122,7 +122,8 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
 
     try
     {
-        int flags = cv::CALIB_RATIONAL_MODEL;
+        // see https://stackoverflow.com/a/75865177
+        int flags = cv::CALIB_RATIONAL_MODEL | cv::CALIB_USE_LU;
         repError = cv::calibrateCamera(all_obj_points, all_img_points, frame_shape, camera_matrix, dist_coeffs, r_vecs, t_vecs, cv::noArray(), cv::noArray(), cv::noArray(), flags);
     }
     catch (...)
@@ -203,7 +204,7 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
         std::vector<cv::Point3f> obj_points;
         std::vector<cv::Point2f> img_points;
 
-        mrcal_point3_t points[(board_width - 1) * (board_height - 1)];
+        std::vector<mrcal_point3_t> points((board_width - 1) * (board_height - 1));
 
         charuco_detector.detectBoard(frame_gray, charuco_corners, charuco_ids, marker_corners, marker_ids);
 
@@ -222,7 +223,7 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
             points[id].z = 1.0f;
         }
 
-        for (int i = 0; i < (sizeof(points) / sizeof(points[0])); i++)
+        for (int i = 0; i < points.size(); i++)
         {
             if (points[i].z != 1.0f)
             {
@@ -232,10 +233,8 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
             }
         }
 
-        std::vector<mrcal_point3_t> points_vector(points, points + (sizeof(points) / sizeof(points[0])));
-
-        frames_rt_toref.push_back(getSeedPose(points_vector.data(), boardSize, imagerSize, square_width, 1000));
-        observation_boards.insert(observation_boards.end(), points_vector.begin(), points_vector.end());
+        frames_rt_toref.push_back(getSeedPose(points.data(), boardSize, imagerSize, square_width, 1000));
+        observation_boards.insert(observation_boards.end(), points.begin(), points.end());
 
         if (show_debug_window)
         {
@@ -250,7 +249,8 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
     }
 
     video_capture.release();
-    cv::destroyAllWindows();
+    if (show_debug_window){
+    cv::destroyAllWindows();}
 
     if (observation_boards.empty())
     {
@@ -361,9 +361,9 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
             frames_rt_toref.push_back(getSeedPose(current_points.data(), boardSize, imagerSize, square_width * 0.0254, 1000));
             observation_boards.insert(observation_boards.end(), current_points.begin(), current_points.end());
         }
-
+        if (show_debug_window){
         cv::imshow("Checkerboard Detection", frame);
-
+        }
         if (cv::waitKey(30) == 'q')
         {
             break;
@@ -371,7 +371,9 @@ int cameracalibration::calibrate(const std::string &input_video, float square_wi
     }
 
     video_capture.release();
+    if (show_debug_window){
     cv::destroyAllWindows();
+    }
 
     if (observation_boards.empty())
     {
