+++ wpical/src/main/native/cpp/fieldcalibration.cpp
@@ -15,7 +15,7 @@
 
 class PoseGraphError {
 public:
-  PoseGraphError(Pose t_ab_observed)
+  explicit PoseGraphError(Pose t_ab_observed)
       : m_t_ab_observed(std::move(t_ab_observed)) {}
 
   template <typename T>
@@ -333,7 +333,8 @@ bool fieldcalibration::process_video_file(
         constraints,
     bool show_debug_window)
 {
-  cv::namedWindow("Processing Frame", cv::WINDOW_NORMAL);
+  if (show_debug_window){
+  cv::namedWindow("Processing Frame", cv::WINDOW_NORMAL);}
   cv::VideoCapture video_input(path);
 
   if (!video_input.isOpened())
@@ -458,7 +459,9 @@ bool fieldcalibration::process_video_file(
   }
 
   video_input.release();
+  if (show_debug_window){
   cv::destroyAllWindows();
+  }
 
   return true;
 }
